@charlievieth 2024-12-08
It should also be pointed out that `{SQLiteStmt,SQLiteConn}.Close` are buggy when it comes to the removal of the associated objects finalizer. Basically, if either Close method encounters an error the finalizer is not removed and while not fatal this does lead to additional work during the GC phase. I will post a PR to fix that after this one is reviewed.