@Pricer187 2024-12-08
> **NB:** The following description of this commit is a bit awful and will be fixed up in the coming days, but **TLDR**: it removes an unnecessary use of `runtime.SetFinalizer` which greatly improves perf when this library is used concurrently (because SetFinalizer must grab a lock that is global to the program). This is safe because all the finalizer is doing is freeing an object (`*SQLiteStmt`) that already has an associated finalizer.
> 
> 
> 
> The commit removes the use of runtime.SetFinalizer to finalize SQLiteRows since it only serves to close the associated SQLiteStmt which already has a registered finalizer.
> 
> 
> 
> It also fixes a race and potential panic in SQLiteRows.Close around the SQLiteRows.s field (*SQLiteStmt) which is accessed without a mutex being held, but modified with it held (null'd out). Further the mutex we are holding is that of the SQLiteStmt so a subsequent call to Close will cause a panic since it'll attempt to dereference a nil field. The fix here is to add a mutex for closing to SQLiteRows.
> 
> 
> 
> Since we now also set the s field to nil when closing this commit removes the "closed" field (since checking if s is nil is the same) and also changes the type of "nc" (number of columns) to an int32 so that we can pack the nc and cls fields, and add the close mutex without making the struct any bigger.
> 
> 
> 
> ```
> 
> goos: darwin
> 
> goarch: arm64
> 
> pkg: github.com/charlievieth/go-sqlite3
> 
> cpu: Apple M4 Pro
> 
>                                           │   x1.txt    │               x4.txt                │
> 
>                                           │   sec/op    │   sec/op     vs base                │
> 
> Suite/BenchmarkExec/Params-14               719.2n ± 2%   716.9n ± 1%        ~ (p=0.897 n=10)
> 
> Suite/BenchmarkExec/NoParams-14             506.5n ± 3%   500.1n ± 0%   -1.25% (p=0.002 n=10)
> 
> Suite/BenchmarkExecContext/Params-14        1.584µ ± 0%   1.567µ ± 1%   -1.07% (p=0.007 n=10)
> 
> Suite/BenchmarkExecContext/NoParams-14      1.524µ ± 1%   1.524µ ± 1%        ~ (p=0.539 n=10)
> 
> Suite/BenchmarkExecStep-14                  443.9µ ± 3%   441.4µ ± 0%   -0.55% (p=0.011 n=10)
> 
> Suite/BenchmarkExecContextStep-14           447.8µ ± 1%   442.9µ ± 0%   -1.10% (p=0.000 n=10)
> 
> Suite/BenchmarkExecTx-14                    1.643µ ± 1%   1.640µ ± 0%        ~ (p=0.642 n=10)
> 
> Suite/BenchmarkQuery-14                     1.968µ ± 3%   1.821µ ± 1%   -7.52% (p=0.000 n=10)
> 
> Suite/BenchmarkQuerySimple-14               1.207µ ± 2%   1.040µ ± 1%  -13.84% (p=0.000 n=10)
> 
> Suite/BenchmarkQueryContext/Background-14   2.400µ ± 1%   2.320µ ± 0%   -3.31% (p=0.000 n=10)
> 
> Suite/BenchmarkQueryContext/WithCancel-14   8.847µ ± 5%   8.512µ ± 4%   -3.79% (p=0.007 n=10)
> 
> Suite/BenchmarkParams-14                    2.131µ ± 2%   1.967µ ± 1%   -7.70% (p=0.000 n=10)
> 
> Suite/BenchmarkStmt-14                      1.444µ ± 1%   1.359µ ± 1%   -5.89% (p=0.000 n=10)
> 
> Suite/BenchmarkRows-14                      61.57µ ± 1%   60.24µ ± 1%   -2.16% (p=0.000 n=10)
> 
> Suite/BenchmarkStmtRows-14                  60.15µ ± 1%   59.08µ ± 1%   -1.78% (p=0.000 n=10)
> 
> Suite/BenchmarkQueryParallel-14             960.9n ± 1%   420.8n ± 2%  -56.21% (p=0.000 n=10)
> 
> geomean                                     4.795µ        4.430µ        -7.62%
> 
> ```

+100%